- Integration test  - tons of bug-fixes

+++ src/main/java/eu/vstoyanov/bitbot/strategy/TradingStrategy.java
@@ -2,8 +2,9 @@ package eu.vstoyanov.bitbot.strategy;
 
 import eu.vstoyanov.bitbot.model.Order;
 
+import java.math.BigDecimal;
 import java.time.Instant;
 
 public interface TradingStrategy {
-    Order tick(Instant timestamp, double price);
+    Order tick(Instant timestamp, BigDecimal price);
 }
